backend/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
# CRITICAL: Fix for passlib conflict
bcrypt==4.0.1